Initial implementation of feature flag pkg
There are more and more experimental changes. This pkg implements
a unified way to control what feature flags (experiment flags) enabled
and provides a way to change some of them in runtime (PATCH request).

From now on all code that considered to be experimental or not desirable by
everyone (e.x. incompatible with graphite-web) should be under a feature flag

Overhead of the implementation is low for single thread - 50ns per request,
Though it's better to replace whole function if the code path is too hot.